Create Nigerian YOR.json transcript
Kudos!, Dear Team leader.
To foster and enable easy interactions for one of Nigeria's most popular tribes to be able to interact and benefit using this app, here is the Yoruba transcript, accurately translated to the Yoruba dialect for our numerous enthusiastic Yoruba Pioneers here in Nigeria. The transcript is Attributed to one of the Nigerian  tribes known as the "Yoruba" tribe, while it is  acronymed as 'YOR'.